+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # WTIntExtensions
 
+## 1.0.3
+
+- Adds an error type, `WTIntExtensionsError`
+- The static function `randomNonZero(::)` now throws an error when both arguments are zero.
+
 ## 1.0.2
 
 - Aligns versioning with other of my libraries.